+++ src/convtools/__init__.py
@@ -4,4 +4,4 @@ from ._conversion import conversion  # noqa: F401
 from ._dt import DateGrid, DateTimeGrid
 
 
-__version__ = "1.15.0"
+__version__ = "1.16.0"